refactor: extract runtime WP-CLI bridge helpers
